Refactoring of window package.
