+++ src/main/java/Graph/Node.java
@@ -1,4 +1,4 @@
-package Parsing;
+package Graph;
 
 /**
  * Created by 101010.
