+++ src/main/java/Window/GraphScene.java
@@ -1,0 +1,9 @@
+package Window;
+
+import javafx.scene.layout.Pane;
+
+/**
+ * Created by Michael on 5/12/2017.
+ */
+public class GraphScene extends Pane {
+}
