+++ src/main/java/Window/Window.java
@@ -1,12 +1,10 @@
 package Window;
 
 import javafx.application.Application;
-import javafx.geometry.Pos;
 import javafx.scene.Scene;
 import javafx.scene.control.Menu;
 import javafx.scene.control.MenuBar;
 import javafx.scene.layout.BorderPane;
-import javafx.scene.layout.VBox;
 import javafx.stage.Stage;
 
 /**
@@ -19,16 +17,16 @@ public class Window extends Application{
 
         BorderPane mainPane = new BorderPane();
         mainPane.setMinSize(1000, 500);
-        mainPane.setLeft(createVBoxLeft());
-        mainPane.setRight(createVBoxRight());
+        mainPane.setRight(new InfoScreen());
         mainPane.setTop(createMenuBar());
+
 
         //Creating a scene object
         Scene scene = new Scene(mainPane);
         scene.getStylesheets().add("layoutstyles.css");
 
         //Setting title to the Stage
-        stage.setTitle("Registration Form");
+        stage.setTitle("Main window");
 
         //Adding scene to the stage
         stage.setScene(scene);
@@ -37,21 +35,6 @@ public class Window extends Application{
         stage.show();
     }
 
-    public VBox createVBoxLeft() {
-        VBox box = new VBox();
-        box.getStyleClass().addAll("pane", "vbox");
-        box.setAlignment(Pos.CENTER);
-        box.setMinSize(200, 200);
-        return box;
-    }
-
-    public VBox createVBoxRight() {
-        VBox box = new VBox();
-        box.getStyleClass().addAll("pane", "vbox");
-        box.setAlignment(Pos.CENTER);
-        box.setMinSize(200, 200);
-        return box;
-    }
 
     public MenuBar createMenuBar() {
         MenuBar menuBar = new MenuBar();
